FIX: MatchList component에서 MatchCard를 분리하여 별도 component로 작성
MatchList의 크기가 커짐에 따라 별도로 분리하는 것이 작업에 용이할 것으로 판단.
